fix: split files into separate folder, update imports
